fix short saga correlation properties (#7788) (#7792)
* fix short saga correlation properties

* Refactor saga header mapping tests to cover all the cases once and for-all

* Assert value directly instead of string

* Remove redundant body

* Better coverage for property finder and explanation why accessors don't require more permutation tests

* Comment

---------

+++ src/NServiceBus.Core/Sagas/HeaderPropertySagaFinder.cs
@@ -81,9 +81,9 @@ class HeaderPropertySagaFinder<TSagaData>(string headerName, string correlationP
             _ when correlationPropertyType == typeof(uint) =>
                 uint.Parse(headerValue, CultureInfo.InvariantCulture),
             _ when correlationPropertyType == typeof(short) =>
-                int.Parse(headerValue, CultureInfo.InvariantCulture),
+                short.Parse(headerValue, CultureInfo.InvariantCulture),
             _ when correlationPropertyType == typeof(ushort) =>
-                uint.Parse(headerValue, CultureInfo.InvariantCulture),
+                ushort.Parse(headerValue, CultureInfo.InvariantCulture),
             _ => ThrowUnsupportedExceptionForCorrelationPropertyType(correlationPropertyType)
         };
     }
